refactor: amend ZWOASICameraLib location to lib in zwoasi module
refactor: amend ZWOASICameraLib location to lib in zwoasi module
